Move interop class and interface to services folder
