--- src/FullCalendarBlazor/JsInterops/IFullCalendarJsInterop.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace FullCalendarBlazor.JsInterops
-{
-    public interface IFullCalendarJsInterop
-    {
-        ValueTask Render(string elementId);
-    }
-}
