+++ src/FullCalendarBlazor/Services/IJSRuntimeService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using FullCalendarBlazor.Models;
+
+namespace FullCalendarBlazor.Services
+{
+    public interface IJSRuntimeService
+    {
+        ValueTask Render(string elementId, FullCalendarData data);
+    }
+}
